moved time format function to a specific file
